# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      summary: Get a single lesson
      responses:
        "200":
          description: Lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
        "404":
          description: Not found
  /lessons/{lessonId}/quiz:
    get:
      operationId: getLessonQuiz
      tags: [lessons]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      summary: Get quiz for a lesson
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /lessons/{lessonId}/quiz/submit:
    post:
      operationId: submitQuiz
      tags: [lessons]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      summary: Submit quiz answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"

  # PROGRESS
...
  /progress/complete:
    post:
      operationId: markLessonComplete
      tags: [progress]
      summary: Mark a lesson as complete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonCompleteInput"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseProgress"
  /progress/certificate/{courseId}:
    get:
      operationId: getCertificate
      tags: [certificates]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      summary: Get certificate data for completed course
      responses:
        "200":
          description: Certificate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
        "403":
          description: Course not completed

  # ACCESS CODES
...
components:
  schemas:
...
    LessonDetail:
      type: object
      required: [id, courseId, title, titleAr, titleSo, order, duration, isLocked, content, contentAr, contentSo]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        titleSo:
          type: string
        order:
          type: integer
        duration:
          type: string
        isLocked:
          type: boolean
        hasQuiz:
          type: boolean
        content:
          type: string
        contentAr:
          type: string
        contentSo:
          type: string
        videoUrl:
          type: ["string", "null"]

    # QUIZ
    Quiz:
      type: object
      required: [id, lessonId, questions]
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
    QuizQuestion:
      type: object
      required: [id, question, options]
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
    QuizSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [questionId, answer]
            properties:
              questionId:
                type: integer
              answer:
                type: integer
    QuizResult:
      type: object
      required: [score, total, passed]
      properties:
        score:
          type: integer
        total:
          type: integer
        passed:
          type: boolean

    # PROGRESS
    CourseProgress:
      type: object
      required: [courseId, userId, completedLessons, totalLessons, percentComplete, isCompleted]
      properties:
        courseId:
          type: integer
        userId:
          type: integer
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        percentComplete:
          type: number
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
    LessonCompleteInput:
      type: object
      required: [lessonId]
      properties:
        lessonId:
          type: integer

    # CERTIFICATE
    Certificate:
      type: object
      required: [userId, userName, courseId, courseName, completedAt]
      properties:
        userId:
          type: integer
        userName:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        completedAt:
          type: string
          format: date-time

    # ACCESS CODES
...