# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from './quizQuestion';

export interface Quiz {
  id: number;
  lessonId: number;
  questions: QuizQuestion[];
}